chanstate: move commitment value types
Move ChannelCommitment and HTLC into chanstate so upcoming store facets
can name commitment state without importing channeldb.

Leave the KV serialization helpers in channeldb and keep aliases for
existing call sites. This preserves the current disk format and keeps
backend-specific persistence code out of chanstate for now.
